langgraph_c4_generator/with_ui/requirements_chatbot.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
langgraph>=0.0.20
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
tiktoken>=0.5.0
